research_pipeline/pyproject.toml
[build-system]
requires = ["setuptools>=68.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "research-pipeline"
version = "0.1.0"
description = "Research pipeline — conversational retrieval over ingested documents"
requires-python = ">=3.12"
dependencies = [
    "fastapi>=0.115",
    "uvicorn>=0.30",
    "pydantic>=2.0",
    "openai>=1.30",
    "sqlalchemy>=2.0",
    "psycopg2-binary>=2.9",
    "pgvector>=0.3",
    "python-dotenv>=1.0",
    "requests>=2.31",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0",
    "pytest-cov>=5.0",
    "black>=24.0",
    "flake8>=7.0",
    "pylint>=3.0",
]

[tool.black]
line-length = 99
target-version = ["py312"]

[tool.pylint.format]
max-line-length = 99

[tool.pylint."messages control"]
disable = []

[tool.pytest.ini_options]
testpaths = ["tests"]
